sales_chatbot/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-experimental
google-generativeai
mysql-connector-python
python-dotenv
fastapi
uvicorn
pydantic
requests